==> scr/00_libraries.R <==
renv::install("Seurat")
renv::install("R.utils")
renv::install("satijalab/seurat-wrappers")
renv::install("hdf5r")
renv::install("tidyverse")

renv::install("dittoSeq")
renv::install("clustree")
renv::install("pals")
renv::install("magick")

renv::install("data.tree")
renv::install("HGNChelper")
renv::install("openxlsx")


renv::install("bioc::scater")
renv::install(c("robustbase","harmony"))
renv::install("jokergoo/ComplexHeatmap")
renv::install("homologene")
renv::install("immunogenomics/presto")

renv::install("GGally")
